+++ lib/Text/TaggerQL.hs
@@ -29,80 +29,54 @@ module Text.TaggerQL (
   combinableSentenceResultSet,
 ) where
 
-import Control.Monad (void, (<=<))
-import Data.HashSet (HashSet)
-import qualified Data.HashSet as HS
-import Data.Hashable (Hashable)
+import Control.Monad (void)
+import Control.Monad.Trans.Class (MonadTrans (lift))
+import Control.Monad.Trans.Reader (
+  ReaderT (runReaderT),
+  ask,
+  asks,
+  local,
+ )
 import qualified Data.List as L
 import Data.List.NonEmpty (NonEmpty ((:|)))
 import Data.Maybe (fromJust)
 import Data.String (IsString (..))
-import Data.Tagger (
-  QueryCriteria (
-    DescriptorCriteria,
-    FilePatternCriteria,
-    MetaDescriptorCriteria,
-    UntaggedCriteria
-  ),
-  SetOp (..),
- )
 import Data.Text (Text)
 import qualified Data.Text as T
 import Database.Tagger.Query (
-  flatQueryForFileByTagDescriptorPattern,
-  flatQueryForFileOnMetaRelationPattern,
   insertTags,
   queryForDescriptorByPattern,
-  queryForFileByPattern,
   queryForTagByFileKeyAndDescriptorPatternAndNullSubTagOf,
   queryForTagBySubTagTriple,
-  queryForUntaggedFiles,
  )
 import Database.Tagger.Type (
   Descriptor (descriptorId),
   File,
   RecordKey,
   Tag (tagId),
+  TaggedConnection,
  )
 import System.IO (hPrint, stderr)
 import Text.TaggerQL.AST (
   ComplexTerm (..),
-  Request (Request),
   Sentence (Sentence),
-  SentenceSet (..),
-  SentenceTree (..),
   SimpleTerm (..),
   Term (Term),
   TermTree (..),
   complexTermNode,
  )
-import Text.TaggerQL.Engine.QueryEngine
-import Text.TaggerQL.Engine.QueryEngine.Type
+import Text.TaggerQL.Engine.QueryEngine (
+  CombinableSentenceResult,
+  TaggerQLQuery (..),
+  combinableSentenceResultSet,
+  combinableSentenceResultSetOp,
+  queryRequest,
+  taggerQL,
+ )
 import Text.TaggerQL.Parser.Internal (
   parse,
-  requestParser,
   sentenceParser,
  )
-
-data CombinableSentenceResult
-  = CombinableSentenceResult SetOp (HashSet File)
-  deriving (Show, Eq)
-
-combinableSentenceResultSetOp :: CombinableSentenceResult -> SetOp
-combinableSentenceResultSetOp (CombinableSentenceResult so _) = so
-
-combinableSentenceResultSet :: CombinableSentenceResult -> HashSet File
-combinableSentenceResultSet (CombinableSentenceResult _ s) = s
-
-{- |
- newtype wrapper for 'Text`
-
- is an instance of IsString
--}
-newtype TaggerQLQuery = TaggerQLQuery Text deriving (Show, Eq)
-
-instance IsString TaggerQLQuery where
-  fromString = TaggerQLQuery . T.pack
 
 {- |
  newtype wrapper for 'Text`
@@ -114,139 +88,9 @@ newtype TaggerQLTagStmnt = TaggerQLTagStmnt Text deriving (Show, Eq)
 instance IsString TaggerQLTagStmnt where
   fromString = TaggerQLTagStmnt . T.pack
 
-newtype TermResult = TermResult {termResult :: HashSet File} deriving (Show, Eq)
-
 newtype TermTag = TermTag (Term Text) deriving (Show, Eq)
 
 newtype TermSubTag = TermSubTag (Term Text) deriving (Show, Eq)
-
-{- |
- Run a 'TaggerQLQuery` on a connection.
--}
-taggerQL :: TaggerQLQuery -> TaggedConnection -> IO (HashSet File)
-taggerQL (TaggerQLQuery q) tc = do
-  let parseResult = parse requestParser "TaggerQL" q
-  either
-    (const (return HS.empty) <=< hPrint stderr)
-    (fmap combinableSentenceResultSet . queryRequest tc)
-    parseResult
-
-{- |
- Run a query given the 'Request` AST of the result of parsing a 'TaggerQLQuery`
-
- Can be used to programmatically generate queries using TaggerQL alongside the
- Text.TaggerQL.AST module to create the structure.
--}
-queryRequest ::
-  TaggedConnection ->
-  Request Text ->
-  IO CombinableSentenceResult
-queryRequest tc (Request strs) = combineSentences <$> mapM (querySentenceTree tc) strs
-
-querySentenceTree ::
-  TaggedConnection ->
-  SentenceTree Text ->
-  IO CombinableSentenceResult
-querySentenceTree tc tr =
-  case tr of
-    SentenceNode ss -> querySentenceSet tc ss
-    SentenceBranch so sss -> do
-      (CombinableSentenceResult _ sentenceResults) <-
-        combineSentences <$> mapM (querySentenceTree tc) sss
-      return $ CombinableSentenceResult so sentenceResults
-
-querySentenceSet ::
-  TaggedConnection ->
-  SentenceSet Text ->
-  IO CombinableSentenceResult
-querySentenceSet tc (CombinableSentence so s) =
-  CombinableSentenceResult so . termResult <$> querySentence tc s
-
-{- |
- All 'TermTree`s in a 'Sentence` are intersected with each other.
--}
-querySentence ::
-  TaggedConnection ->
-  Sentence Text ->
-  IO TermResult
-querySentence tc (Sentence tts) =
-  intersectTermResults
-    <$> mapM (queryTermTree tc) tts
-
-{- |
- Return 'Nothing` if the given 'TermTree` is Complex and Bottom.
--}
-queryTermTree ::
-  TaggedConnection ->
-  TermTree Text ->
-  IO TermResult
-queryTermTree tc tt =
-  case tt of
-    Simple st -> querySimpleTerm tc st
-    Complex ct ->
-      TermResult
-        <$> runReaderT (queryComplexTermTopLevel ct) (QueryEnv mempty tc)
-
-querySimpleTerm ::
-  TaggedConnection ->
-  SimpleTerm Text ->
-  IO TermResult
-querySimpleTerm tc (SimpleTerm (Term qc p)) =
-  case qc of
-    DescriptorCriteria ->
-      TermResult . HS.fromList <$> flatQueryForFileByTagDescriptorPattern p tc
-    MetaDescriptorCriteria ->
-      TermResult . HS.fromList <$> flatQueryForFileOnMetaRelationPattern p tc
-    FilePatternCriteria ->
-      TermResult . HS.fromList <$> queryForFileByPattern p tc
-    UntaggedCriteria ->
-      TermResult . HS.fromList <$> queryForUntaggedFiles tc
-
-{- |
- Should perform an associative strict intersection of the given 'TermResult`s.
--}
-intersectTermResults :: [TermResult] -> TermResult
-intersectTermResults [] = emptyTermResult
-intersectTermResults trs = L.foldl1' intersectTermResult trs
-
-intersectTermResult :: TermResult -> TermResult -> TermResult
-intersectTermResult (TermResult x) (TermResult y) = TermResult $ HS.intersection x y
-
-combineSentences :: [CombinableSentenceResult] -> CombinableSentenceResult
-combineSentences [] = emptyCombinableSentenceResult
-combineSentences cs = L.foldl1' combineSentence cs
-
-{- |
- Left-associative combination of the results of a combinable sentence query.
--}
-combineSentence ::
-  CombinableSentenceResult ->
-  CombinableSentenceResult ->
-  CombinableSentenceResult
-combineSentence
-  (CombinableSentenceResult sox sx)
-  (CombinableSentenceResult soy sy) = CombinableSentenceResult sox $ combFun soy sx sy
-
-{- |
- Return an empty result with a default 'SetOp` of Union.
--}
-emptyCombinableSentenceResult :: CombinableSentenceResult
-emptyCombinableSentenceResult = CombinableSentenceResult Union HS.empty
-
-emptyTermResult :: TermResult
-emptyTermResult = TermResult HS.empty
-
-combFun ::
-  Hashable a =>
-  SetOp ->
-  HashSet a ->
-  HashSet a ->
-  HashSet a
-combFun so =
-  case so of
-    Union -> HS.union
-    Intersect -> HS.intersection
-    Difference -> HS.difference
 
 type TaggingEnvironment = (TaggedConnection, [RecordKey Tag])